update: thesis version

+++ dataRawProcess/audio_processing/rating_audio.py
@@ -14,7 +14,7 @@ def run_predict(audio_folder_path, playlist_name, episode):
 
     os.makedirs(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}', exist_ok=True)
     if os.path.exists(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}_NISQA_results.csv'):
-        logger.info(f"File {episode}_wvmos.csv already exists, skipping...")
+        logger.info(f"File {episode}_NISQA_results.csv already exists, skipping...")
         return
     # os.makedirs(f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}/{episode}', exist_ok=True)
     # command = f"python run_predict.py --mode predict_dir --pretrained_model weights/nisqa.tar --data_dir '{audio_folder_path}' --num_workers 8 --bs 30 --output_dir /home4/tuanlha/EXpressiveTTS/dataRawProcess/05_data_extract/{playlist_name}"
@@ -35,7 +35,7 @@ def limit_cpu_for_diarization():
     """Giới hạn process và các subprocess chỉ chạy trên core 0 và 1"""
     try:
         p = psutil.Process(os.getpid())
-        p.cpu_affinity([i for i in range(34,40)])  # Chỉ dùng core 0 và 1
+        p.cpu_affinity([i for i in range(32,40)])  # Chỉ dùng core 0 và 1
         print(f"Process {os.getpid()} bị giới hạn trên core: {p.cpu_affinity()}")
     except Exception as e:
         print(f"Không thể thiết lập cpu_affinity: {e}")
@@ -43,7 +43,7 @@ def limit_cpu_for_diarization():
 def rating_audio(args, cfg):
 
     playlist_name = args.playlist_name
-    episode_list = sorted(os.listdir(os.path.join(args.data_path, args.playlist_name)))
+    episode_list = sorted(os.listdir(os.path.join('./00_standardization', args.playlist_name)))
     episode_name = [os.path.basename(ep).rsplit('.', 1)[0] for ep in episode_list]
 
     audio_folder_path = f'/home4/tuanlha/EXpressiveTTS/dataRawProcess/04_denoise/{playlist_name}'
@@ -76,25 +76,30 @@ def rating_audio(args, cfg):
 
 def process_filter_by_rating(args, episode):
     csv_path = f'./05_data_extract/{args.playlist_name}/'
-    nisqa_path = os.path.join(csv_path, f'{episode}_NISQA_results.csv')
-    wvmos_path = os.path.join(csv_path, f'{episode}_wvmos.csv')
-    df_1 = pd.read_csv(nisqa_path)
-    df_2 = pd.read_csv(wvmos_path)
     
-    # From df_1, remove audio which MOS < 3.5
-    df_1 = df_1[df_1['mos_pred'] < 3.5]['deg']
-    df_1 = df_1.tolist()
+    try:
+        nisqa_path = os.path.join(csv_path, f'{episode}_NISQA_results.csv')
+        df_1 = pd.read_csv(nisqa_path)
+        df_1 = df_1[df_1['mos_pred'] < 3.5]['deg']
+        df_1 = df_1.tolist()
+    except Exception as e:
+        logger.error(f"Error reading NISQA results for episode {episode}: {e}")
+        df_1 = None
+
+    wvmos_path = os.path.join(csv_path, f'{episode}_wvmos.csv')
+    df_2 = pd.read_csv(wvmos_path)
     df_2 = df_2[df_2['score']<2.75]['audio_path']
     df_2 = df_2.tolist()
 
     folder_path = os.path.join('./04_denoise', args.playlist_name, episode)
 
-    for file in df_1:
-        file = os.path.join(folder_path, file)
-        if os.path.exists(file):
-            os.remove(file)
-        else:
-            print(f"File {file} does not exist")
+    if df_1 is not None:
+        for file in df_1:
+            file = os.path.join(folder_path, file)
+            if os.path.exists(file):
+                os.remove(file)
+            else:
+                print(f"File {file} does not exist")
     for file in df_2:
         file = os.path.join(folder_path, file)
         if os.path.exists(file):
@@ -105,7 +110,7 @@ def process_filter_by_rating(args, episode):
 
 def filter_by_rating(args, cfg):
     playlist_name = args.playlist_name
-    episode_list = sorted(os.listdir(os.path.join(args.data_path, args.playlist_name)))
+    episode_list = sorted(os.listdir(os.path.join('./00_standardization', args.playlist_name)))
     episode_name = [os.path.basename(ep).rsplit('.', 1)[0] for ep in episode_list]
     
     with ThreadPoolExecutor(max_workers=4) as executor:
